Support both adding and removing artist and album favorites
Currently, the "Add to favorites" function in both the artist and album menus is unaware of the current favorite status. This change will allow toggling it by displaying either "Add to" or "Remove from" favorites, depending on the current state. Also, in order to eliminate confusion with LMS favorites, "favorites" is changed to "Qobuz favorites" in the message text.