Group unstreamable tracks with the others in their work.
For some reason, Qobuz does not populate the composer tag for classical tracks that are tagged as unavailable for streaming. This prevents them from being included in the 'work' object to which they belong in the plugin, which is keyed on composer and work. Instead, they are added to separate works, keyed on the work name only, which is confusing. Fortunately there are not a whole lot of releases containing these tracks, but enough to bother someone with a low threshold of annoyance such as myself.  ;-)  This change groups these tracks with the others in same work.