Allow orphaned album favorites to be removed.
This change allows album favorites whose associated album id's are not found in the Qobuz library to be removed via the Album Favorites menu, saving the user from having to perform the task in the Qobuz app.

(Please check my Google Translate-generated message translations.)

An additional minor change prevents unstreamable tracks that are not also tagged as sampleable from being added to the player queue.