Change how unstreamable albums are handled
Allow unstreamable albums to be clicked through to the tracklist menu, where any relevant messages are then displayed. The "*" is still prepended to the album title in the album list, indicating that the title is either not yet released or marked as unstreamable. The details are then shown on the track list page. In the case of not-yet-released titles, the date when it is scheduled for release is displayed.