Test the release date accurately and efficiently
Since the releaase_at tag is based on the French time zone (CEST), it cannot be used accurately in other time zones close to the release time when determining if the album is released yet. In those cases only (within +/- 14 hours of the release time), we will use the release_date_stream tag to make the determination.

Also remove the code in _trackItem() that supposedly checks the release time of each track but does nothing, since there is no released_at attribute in the track object and checking it again for each track's album would be redundant.