Improve handling and display of release dates
These changes simplify how the release date and year are calculated and displayed, removing discrepancies  between the YEAR tag and the release date in time zones east of France for albums released on January 1, among other things. Note: It would be nice to have a version of Slim::Utils:DateTime::shortDateF() that took a localtime structure instead of an epoch time as an argument.

The change to strings.txt corrects the statement that sample tracks are 60 seconds. They are now 30 seconds.

Albums that are not yet released but streamable due to having one or more pre-released playable tracks are no longer filtered out in Common.pm when the user has chosen not to play 30-second sample tracks, allowing those albums to be shown and the pre-release tracks to be played.