Allow grouping of artist releases by release type.
This is a change I have been wanting to make for a long time. It adds an option to group releases for an artist by release type, as defined by the iTunes and Spotify standards. With the option in effect, each release is classified as either an Album, EP or Single and grouped under one of those three headings. The order of the the groups is not configurable.  However, the releases within each group will be sorted as specified in the Sort Artist Albums option. For comparison purposes, and to avoid confusion with Qobuz's somewhat random release typing, the Qobuz release type is now displayed in the Album Info page following the Genre, and is labeled as "Qobuz Release Type". All string translations are straight from Google Translate as I am monolingual.