Make Replay Gain aware of RandomPlay plugin settings.
This change allows "Smart gain" to function more accurately with Qobuz tracks when the Random Play plugin is in use. It will use the "Album gain" value when Random Play is in effect and the Mix Type is "Releases" or "Works". For all other Mix Types, "Track Gain" will be used.