Only populate track performer for main artists.
Only populate the Track Artist attribute for performers who are tagged as "MainArtist" or "Main Artist" in the $track->{performers} string. The purpose is to clean up the Track Artist attribute so that both the album artist and the track artist (if any) can be displayed consistently in track lists for all Qobuz releases, whether they come from the plugin or from the library.

Signed-off-by: Sam Y <syahres@gmail.com>